requirements-gpu.txt
tensorflow-gpu==2.5.0rc0
opencv-python==4.1.1.26
lxml
tqdm
absl-py
matplotlib
easydict
pillow
